requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Code quality and formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Development dependencies
python-dotenv==1.0.0

# GraphQL dependencies
strawberry-graphql[fastapi]==0.214.0 